Refactoring: Separate Files
